[Offload] Adds offload-runtime buildbot
This adds a new worker for AMD GPU workloads and uses that worker for
the new offload-runtime buildbot.
That builder is in preparation for landing
https://github.com/llvm/llvm-project/pull/75125.

The builder enables clang and flang frontend together with the runtimes
'offload' and 'openmp' to build host-side OpenMP and GPU offloading
OpenMP.
It runs 'check-openmp' for host-side tests, 'check-clang' and
'check-flang' for frontend/codegen tests and 'check-offload'
for runtime offloading tests.