[flang] Disable OpenMP on flang-aarch64-libcxx bot
Building LLVM with libc++ and OpenMP as a runtime on a glibc based
system causes some hard to fix build failures.

As the point of building against libc++ was to make sure flang works
well with it, adding OpenMP doesn’t help that, and flang OpenMP
tests are already enabled on other flang bots, just disable OpenMP
on flang-aarch64-libcxx.