[OpenMP] Add offload to other buildbots
This is in anticipation of landing the move to offload. We have one
buildbot that already runs the configuration with offload enabled, which
will hopefully turn green once the patch lands.
Once things are fine on that end, we can land this patch to enable the
offload runtime in the other buildbots as well.