Disable PDB generation in stage1 of the clang-cl self-host bot (#195)
PDB generation is resource-intensive and does not perform well for
statically linked unit tests. The stage2 build generates PDBs to
exercise clang-cl's debug info generation abilities, but we generally do
not use the debug info from the stage 1 builds. PDB build errors are
also a common source of builder unreliability and lead to excessive disk
usage, so disable them for now.

Here is a recent example of PDB unreliability:
https://lab.llvm.org/buildbot/#/builders/123/builds/26898
```
... GeneratorTest.cpp: fatal error C1051: program database file, '...\vc140.pdb', has an obsolete format, delete it and recompile
```