[clangbuilder] Add ability to specify cmake args only used for stage2
Right now, extra_cmake_args to getClangCMakeBuildFactory are applied to
both stage1 and stage2. But there are some cases where we would not want
to use the same configs, e.g. flags that might not be understood by the
current host compiler, or enabling ccache only for stage1 (the cache
will never hit for stage2).

Reviewed as part of #188.