Fix ClangBuilder after f59269f1ef7823fe68775d9a2fcc975673654bae (#196)
The way that change was indented meant that the cmake/build/check steps
of stage 2 were only added if extra_stage2_cmake_args was not empty.

This caused any 2 stage builder without those extra options to skip
building stage 2. Which left them green but not doing their
job, or red if they tried to run the test suite using stage 2.