Add a new 'hint' argument for LLVMBuildFactory factory class.
This parameter is used to apply suffixes to the step names when factory is used
as a nested factory for another one. The suffix will be added to the step name
separated by dash symbol.

As example for 'stageX' hint the following step names will be generated for
the factory's steps:
    cmake-cofigure => cmake-configure-stageX
    build => build-<target>-stageX