[zorg][lldb] Add new LLDB metrics bot (#278)
This patch adds a new job to collect LLDB metrics.

This is heavily based on the `debuginfo-statistics` job (but currently
doesn't publish data to LNT).

Currently this job would do the following:
1. Check out the Clang 19.x release and build it
2. Use the LLDB and Clang from the `lldb-cmake-intel` job (not actually
sure if that job publishes the right artifacts at the moment) to run the
`run_lldb_metrics.sh` script.
3. Said script will attach LLDB to Clang/LLDB and run various commands.
Then it will dump the `statistics dump` command to `stdout` (note we
don't do any kind of averaging of these over multiple runs, since the
metrics we care about **should** stable across runs). We also currently
run these test-scenarios through `hyperfine` and dump the timing data.
But maybe for a first attempt this isn't necessary.