[GHA] Add GitHub Action to run the tests for llvm-zorg itself (#282)
This adds a GitHub action to at least run the very minimal lit tests we
have for zorg itself. With appropriate repository settings it should
provide pass/fail for both PRs and commits.

The GH Actions in the main llvm-repos tend to be gated on `if:
github.repository_owner == 'llvm'`, which I've copied across here.

There's a question mark about how to handle
`test/jenkins/test_build.py`. I've marked it XFAIL as a command-line
`lit` option, as the test ends up trying to call `xcrun` which has no
hope of succeeding on Linux. But perhaps the test is fixable, or it
would be better to mark the test itself XFAIL. Either way, I think the
GH Action in the PR is definitely an improvement over the status quo.