Rename BUILDMASTER_TEST to BUILDBOT_TEST
BUILDBOT_TEST is sufficiently descriptive (after all, 'buildbot' is the
package name and the name of the binary invoked to create a buildmaster)
and it's possible upstream may later remove the buildmaster terminology
<https://github.com/buildbot/buildbot/issues/5382> or alternatively we
might move away from it downstream. As such, it seems prudent to use a
name that would be unaffected.