Rename BUILDMASTER_TEST to BUILDBOT_TEST (#301)
BUILDBOT_TEST is sufficiently descriptive (after all, 'buildbot' is the
package name and the name of the binary invoked to create a buildmaster)
and it's possible upstream may later remove the buildmaster terminology
<https://github.com/buildbot/buildbot/issues/5382> or alternatively we
might move away from it downstream. As such, it seems prudent to use a
name that would be unaffected.

BUILDMASTER_TEST was added in #289 and as the docs PR hasn't landed yet, shouldn't have any users relying on it yet.