libc: make new bootstrap builder a dup of runtimes (#332)
Before renaming the existing debian-dbg-runtimes-build to
debian-dbg-bootstrap-build, make sure that the new name will be
recognized so
that there's no downtime of CI coverage due to renaming the buildbot.

Also remove llvm from the explicit project list; that's not necessary
(or may have
been for old hdrgen which has been removed).

Link: https://github.com/llvm/llvm-zorg/pull/325#discussion_r1870167521