[CI] Adjust CPU and memory limits
After switching to n2-highcpu-32 machines, the CPU and memory
requests/limits were left at the same values. This led to pods not
being scheduled or failing due to running out of memory relative to
the requests.

This patch updates the values to reflect the actual machine shapes
that we are running on.