Compile for Neoverse N1 on Linaro's AArch64 bots that are on Mt Jade hardware
This reflects the host hardware, rather than the limited subset of features
on the Cortex-A57.

```
$ lscpu
Architecture:            aarch64
  CPU op-mode(s):        32-bit, 64-bit
<...>
Vendor ID:               ARM
  Model name:            Neoverse-N1
<...>
    Flags:               fp asimd evtstrm aes pmull sha1 sha2 crc32 atomics fphp asimdhp cpuid
                         asimdrdm lrcpc dcpop asimddp
```

I have also confirmed that the Neoverse V1 (Note: V not N) that is
used for Graviton is a superset of these features.

https://github.com/aws/aws-graviton-getting-started/blob/main/c-c%2B%2B.md#enabling-arm-architecture-specific-features

So if we have to move these builds to Graviton 3 or 4 at short notice,
the existing configuration will work.