[CI] Add annotation to prevent eviction during node scale down
Because we are using GKE's autoscaling features, we need to ensure that either
the runner or workflow pod are marked with an annotation prevent scale down from
happening on the node. Otherwise GKE will decide that nodes actively running pods
should be scaled down rather than looking for empty nodes, with the idea being
that applications should be resilient against individual pods being killed,
which we are not.