[CI] Bump service pool node count/node spec
This patch bumps the number of nodes in the service pool from 2 to 3 and
additionally increases the spec of the machines from e2-small to
e2-highcpu-4. We were running into DNS issues due to a downed pod
that we believe was caused by a lack of resources in the service pool.
Adding an additional pod should help mostly with redundancy but also
with capacity in addition to the capacity increase within each node.