[RISCV] Move clang-riscv-rva23-evl-vec-2stage to annotated builder (#374)
The intent is for the rest of the RISE RISC-V builders to follow, but
this starts the process.

Previously the builders used the framework provided by ClangBuilder.py.
As the build has required more customisation (with more to come - such
as building the llvm-test-suite with the stage1 compiler which isn't
support by ClangBuilder) this is requiring more and more changes to
increase configurability of ClangBuilder, but this makes the logic
harder to follow and I think we've reached the limit of what's
reasonable.

The AnnotatedBuilder infrastructure allows the build logic to just be
put in a shell script. The latest version of llvm-zorg is checked out
and the script is executed. This has the additional advantage of
decoupling script updates from needing to wait for buildmaster
redeploys. With that in mind, I've avoided putting any configuration
through options or environment variables defined by the builder
instantiation in builders.py so there's no need to reason about cases
where e.g. a script update was committed but the buildmaster may or may
not have been updated with a new builders.py.

rise-riscv-build.sh is written with the philosophy that "A little
copying is better than a little dependency." We could look to share more
logic, but we'd end up with the same configurability issue as
ClangBuilder. Instead, the script provides only the configurability
needed for this set of builders.

A future improvement would be to make rise-riscv-build.sh more friendly
to running outside of the buildbot environment (e.g. for local testing).
This is left to future work, as it's not a regression vs the current
approach.

This change has been tested locally along the lines of the instructions
I wrote in

<https://llvm.org/docs/HowToAddABuilder.html#testing-a-builder-config-locally>.
Something like the following is sufficient to use a local downstream
testing llvm-zorg branch for the annotated builder scripts temporarily:

```
--- a/zorg/buildbot/builders/AnnotatedBuilder.py
+++ b/zorg/buildbot/builders/AnnotatedBuilder.py
@@ -77,6 +77,8 @@ def getAnnotatedBuildFactory(
     # Check out zorg so we can run the annotator scripts.
     f.addGetSourcecodeForProject(
         name='update-annotated-scripts',
+        repourl='file:///home/asb/llvm-zorg',
+        branch='synced-worktree',
         project='zorg',
         src_dir='llvm-zorg',
         alwaysUseLatest=True)
```