[RISCV] Add llvm-test-suite to RISC-V annotated builders (#376)
As noted in the inline TODO, it may be worth later changing this so the
test suite is run immediately after stage1 as stage1 + llvm-test-suite
can build and run very quickly, provided rapid feedback.

Although the lit-on-qemu mechanism could be used to run the test suite
under qemu-system, we instead run the test suite under qemu-user. This
has the advantage of simplicity - the suite runs well in this mode and
we don't have the same concerns about testing things that may not work
under qemu-user (such as LLDB or sanitizers) as for the main
llvm-project tests.