[CI] Change Windows Machine Shape to 32 CPU
This patch changes the windows machine shape from 64 core machines to 32 core
machines but bumps the maximum number of nodes/runners to 16 (for a total of
512 cores allocated to Windows testing). We are currently running into some
queueing delays on Windows that are higher than desired. Lowering the core
count but increasing the number of nodes will allow higher throughput as there
are quite a few serial steps in the build process so for the vast majority
of some builds, parallelism isn't even being used (like some of the leaf
projects) due to the majority of the time being taken up setting up caches,
CMake configures, and cloning the repository.