[AMDGPU] Change CWD from within tmp directory (#403)
As pointed out in https://github.com/llvm/llvm-zorg/pull/402 the former
solution deleted the CWD from within that directory, rendering the state
quite useless and resulting in errors on execution.

The approach in here creates a temp directory, changes to it, removes
the old CWD, recreates it, and changes back to that directory. I opted
for this solution as, oddly enough, I had some trouble removing the
  enumerated list of files in the buildtree.