[CI] Add BuildKite API secret reference (#400)
The metrics container only logs Github workflows. Since we want to
compare the performance of the GCP based CI to BuildKite, it would be
convenient to have the BuildKite metrics along the GCP metrics in
Grafana.

This change references a newly added GCP secret the metrics container
will be able to use to read BuildKite data.