LLVM_ENABLE_RUNTIMES=flang-rt for flang-runtime-cuda-* (#393)
Add `depends_on_projects=['flang-rt']` to the flang-runtime-cuda-gcc and
flang-runtime-cuda-clang builders. This prepares the removal of the
"projects" build of the flang runtime in
https://github.com/llvm/llvm-project/pull/124126.

Affected builders:

 * flang-runtime-cuda-gcc
This previously only built the runtime using the top-level
CMakeLists.txt in `flang/runtime/CMakeLists.txt`. This is going to be
replaced with the "standalone runtimes build", with the top-level
`runtimes/CMakeLists.txt`. This still needs Flang to successed, hence
replacing with a bootstrap-build where the `FLANG_RT_*` options are
internally forwarded to the runtimes build.

 * flang-runtime-cuda-clang
This is a manual bootstrapping build which first compiles Clang, then
the runtime out-of-tree. This is replaced with a standalone runtimes
build as described above. Because it needs Flang, also adding Flang to
the enabled projects of the stage1 build.

Neither build runs the `check-*` targets, probably due to the lack of
actual CUDA hardware which running the runtime unittests require.

Affected workers:
 * as-builder-7