[lldb] Temporarily remove --arch flag to dotest
We're trying to move the matrix bot to arm64. But specifying the
`--arch=arm64` isn't behaving as we expect. For some reason both
`--arch=arm64` *and* `--arch=x86_64` are passed to the tests, which
might be the reason why they're failing.

Try to remove it for now and see if that changes anything.