[libc][AMDGPU] Add libc to the runtimes list (#419)
It appears that if something is contained in the `depends_on_project`
list but is not explicitly mentioned as an `enable_runtimes`, while also
an explicit `enable_runtimes` list is passed to the unified tree
builder, that component is auto-added as an `LLVM_ENABLE_PROJECTS`. This
is wrong in the case of libc.

Wile we only care about the version being built for AMDGPU, this appears
to be the easiest fix to unblock our libc bot.