Add documentation on resetting ARC
Upgrading/resetting ARC is a somewhat involved process that cannot be done with
just terraform apply with an updated configuration. This patch adds some
documentation on the exact steps to perform, some of the motivation behind
them, and some tips and tricks that I have developed while hacking on the
cluster.

Reviewers: lnihlen, Keenuts, gburgessiv, cmtice

Reviewed By: lnihlen

Pull Request: https://github.com/llvm/llvm-zorg/pull/420