[RISCV] Add "gauntlet" bot for rapidly testing a range of configs with llvm-test-suite
More so than the other RISC-V builders, this is optimised for getting some
testing feedback as quickly as possible. It isn't intended to replace
any of those, meaning we can rely on them still running for higher
coverage and higher emulation fidelity. The idea of this bot is to
help with the steps we typically end up taking if there's an unexpected
bot failure:
* Figure out if it's easily reproducible on a simple test (especially
  true if there was an issue in a stage2 native build).
* Determine if it affects all or only some configurations.
* Narrow down exactly which patch caused it (helped here due to more
  rapid cycle time of this bot).

It does this by doing a simple stage1 build, then cross-compiling the
llvm-test-suite in multiple configurations and running with qemu-user.
It takes about 25 minutes to run (I masked a couple of 'long pole' tests
that take a lot longer than others).

If we had "gatekeeper"
<https://discourse.llvm.org/t/rfc-introduce-gate-keeper-builders-to-reduce-notification-noise-from-long-running-bots/67931>
bots this would be an ideal candidate - to be run as soon as a fast x86
bot determines that there's failure that's not RISC-V specific. In the
absence of support for that setup, we do it back to front - run the
llvm-test-suite across the configurations of interest. If there were any
failures, then do 'check-all' for the x86 clang compiler so that we can
at least flag in the logs if it appears this likely an issue that isn't
RISC-V specific.