Add repeat and sleep for STEP_FAILURE emission
It doesn't work reliably without it (see
zorg/buildbot/builders/sanitizers/buildbot_functions.sh which ran into
the same problem). It's not clear what the problem is, but a delay alone
is not enough (e.g. printing and sleeping for a minute), the combination
of printing the failure string multiple times and waiting after each one
seems necessary.