[lldb][sanitized] Don't overwrite TARGETS_TO_BUILD
We moved the sanitized bot to arm64 but the LLVM we were building was
only X86-enabled. So when the freshly built Clang was used to compile
the runtimes it failed on arm64.

This was happening because we were overwriting the TARGETS_TO_BUILD for
the sanitized variant. There's no need to overwrite it. We should just
build for X86 *AND* AArch64, which would get us some more LLDB test coverage too.