Clean build directory in cmake monolithic windows (#438)
https://discourse.llvm.org/t/possible-issue-with-premerge-monolithic-windows/86070
showed some issues due to a dirty cache. We should be cleaning the build
directory after each build and letting sccache handle all of the caching

This is for the old infrastructure which should be turned down in a
couple weeks, but it's a simple enough patch in the mean time.