[CI] Misc fixes
This makes a couple misc fixes needed for setting up the new cluster:
* Added a node locations option for the service nodes so that they do not
  get replicated per zone in regional clusters.
* Make the k8s secrets for the runner scale sets depend on the namespace
  resources so that we do not end up in an error caused by a race condition
  where TF tries to create the secrets in a nonexistent namespace.

This is not actually enough to get the second cluster running. Trying to
startup ARC causes failures as we are currently not specifying a runner
group and the runner scale sets need to be placed under different runner
groups.