[CI] Specify runner group per cluster
This patch specifies the runner group that a runner scale set is associated
with specifically per cluster rather than just using the default runner group.
This is a requirement for setting up a HA cluster. The runner scale sets need
to be named the same but in different groups so that they do not conflict.