[CI] Remove Buildkite Token from metrics container
Now that buildkite has been sunsetted, we no longer need to keep track of
the Buildkite workflows. https://github.com/llvm/llvm-project/pull/143049 does
this on the metrics container side. This patch adjusts the deployment to
those changes. Note that patch needs to land first before this can be
applied.