Linaro / Windows: Use LLVM_CCACHE_BUILD instead of CMAKE_CXX_COMPILER_LAUNCHER
There's nothing wrong with the latter - except that LLVM_CCACHE_BUILD
knows that it should not be enabling pre-compiled headers on Windows.

See https://github.com/llvm/llvm-project/pull/136856

We had a contributor report an issue with this, and it only got fixed
when someone randomly managed to update the right files for an unrelated
change.

This sounds like a known issue, and if we use LLVM_CCACHE_BUILD,
precompiled headers will be disabled. This is fine, our bots weren't
relying on this anyway.

As far as I could tell we don't customise the ccache install at all
either, so the defaults are all fine for us.

I've applied this to all 4 Windows bots:
clang-arm64-windows-msvc
clang-arm64-windows-msvc-2stage
flang-arm64-windows-msvc
lldb-aarch64-windows

The problem is with flang specifically, but we might as well standardise
across them all.