Using skip markers in nfc setup
If the NFC check setup fails to execute for any reason, we still want to run
tests unconditionally. To implement this, we now use skip markers: tests run
when these markers are not present.

If the current and previous llvm-bolt binaries are identical, we create two
skip markers for in-tree and out-of-tree tests. If there are other BOLT source
changes, we delete the in-tree skip marker so those test Can proceed.