Upstream commit metrics scraping script and container definition (#483)
This upstreams a script for daily scraping of
[llvm/llvm-project](https://github.com/llvm/llvm-project) for new
commits. The scraped commits are queried against [GitHub Archive
BigQuery](https://www.gharchive.org/#bigquery) to determine useful
metrics such as how many commits are submitted with or without a pull
request and how many of those pull requests are reviewed or approved.

This container will run in the cluster and will push information to a
Grafana dashboard similar to
https://llvm.grafana.net/public-dashboards/21c6e0a7cdd14651a90e118df46be4cc