[BOLT] Add sanity check for frozen llvm-bolt
Some patches can cause the llvm-bolt binary to hang, which stalls or
fails the test pipeline.

Add a simple sanity check that runs:
```
llvm-bolt --version
```

with a 30-second timeout. If the command does not complete in time,
flunk the build.

Set maxTime for nfc-check-validation and reduce the number of lit
workers for in-tree tests.