[CI] Version pin providers
This patch version pins providers and bumps the google provider to
v6.43.0. We need to bump the Google provider so that we can set the windows OS
version on windows node pools for the switch to server 2022. This also pins
some other providers as hashicorp/helm in particular has breaking changes in v3
which terraform init -upgrade will install before this patch.

Reviewers: dschuff, lnihlen, gburgessiv, Keenuts, cmtice

Reviewed By: cmtice

Pull Request: https://github.com/llvm/llvm-zorg/pull/497