Fix out-of-tree Flang builders (#500)
Currently, when Flang is built out-of-tree, it fails to find
compiler-rt when linking programs. This issue was reported in
llvm/llvm-project#147738 and a fix for it is being discussed.

Meanwhile, there is a workaround available, discussed in #482, that
consists in using an installed flang instead of the one in the
build directory. This PR changes Flang builder to use this
workaround.