[CI] Update documentation after libcxx runner sets
There are a couple places where we need to update the documentation to reflect
changes made due to the introduction of the libc++ runner sets. I did a cursory
glance through the documentation and updated everything where I saw something
was incorrect, but this patch might not be exhaustive.

Reviewers: lnihlen, gburgessiv, dschuff, Keenuts, cmtice

Reviewed By: cmtice

Pull Request: https://github.com/llvm/llvm-zorg/pull/510