[CI] Setup Buckets for Sccache
This patch sets up GCS Buckets for sccache remote caching. We set up one for
Linux and one for Windows in each cluster. This creates some separation which
might help slightly with security. We want to have the buckets in the same
region as the machines to ensure uploads/downloads are as low latency as
possible which is helped significantly when the compute and storage are
colocated.

Reviewers: dschuff, Keenuts, cmtice, lnihlen, gburgessiv

Reviewed By: cmtice

Pull Request: https://github.com/llvm/llvm-zorg/pull/505