[CI] Add gcs_bucket_location variable to gke_cluster module
This patch introduces a new gcs_bucket_location variable. Without this we
cannot deploy the GCS buckets in the us-central1 region as the cluster is
a zonal cluster in us-central1-a. Buckets have to be regional and cannot
be specific to a zone, so we would get API errors when running
terraform apply.