[CI] Add namespace names as parameters to premerge_resources
This enables a bit of simplification as we can remove some strings where they
really should be variables. This will also help in a future patch where we also
need to access the namespace name inside the gke_cluster module as it will
enable having a common definition to ensure that they do not differ.

Reviewers: Keenuts, dschuff, gburgessiv, cmtice, lnihlen

Reviewed By: cmtice

Pull Request: https://github.com/llvm/llvm-zorg/pull/507