[CI] Remove need for operational metrics persistent storage (#501)
Originally, the script for scraping LLVM commit info required persistent
storage to keep track of the last commit we've seen. This was to prevent
any overlap in our scrapes and avoid processing the same commits
multiple times.

However, currently the script will only ever scrape a full day at a
time, so there should never be any overlap. We no longer need persistent
storage to maintain the most recently viewed commits since we don't
expect the CronJob to be executed twice in the same day.