Merge clang-armv7-lnt into clang-armv7-2stage (#520)
This removes the single stage test suite bot, then adds stage 1 check
and final test suite run to the 2 stage bot.

Which is the equivalent but 1 fewer worker. It'll slow down the 2 stage
builds a bit but there's no a lot of activity in the 32-bit Arm area
anyway, so I don't think it's a problem.