[CI] Add maintenance window to premerge clusters (#522)
This patch adds an explicit maintenance window to the premerge clusters.
This is in response to some control plane upgrades that we recieved
notice of that will prevent access to the k8s control plane for ~15
minutes which means we will not be able to start new jobs. We should
have this anyways though as the current node upgrade strategy also
breaks jobs.