[CI] Bump maintenance window end date
This patch bumps the maintenance window end date so we can actually apply the
changes. GKE requires that there be at least 48 hours of maintenance time
available in >=4h blocks for every 32 day segment. The original rules did not
meet this. These changes allow us to deploy and should not sacrifice too much
in terms of ensuring we minimize commit traffic during the maintenance
periods.