[CI] Increase resource allocation for operational metrics cronjob (#524)
The cronjob for scraping llvm-project commit data consistently fails due
to running out of memory. `git clone
https://github.com/llvm/llvm-project.git` consistently uses around 1.5
GiB of RAM, which is well above both the request and limit currently set
for the cronjob. Thus, the job is repeatedly killed more memory than is
allocated.

Increasing the resource allocation should get this cronjob running as
intended.