[CI] Validate scraped push commits via GitHub API (#514)
As GitHubArchive BigQuery is known to be lossy, it is likely that we
currently overestimate the number of commits made to llvm/llvm-project
without an associated pull request.

To remedy this, we can make calls to the [GitHub Event
API](https://docs.github.com/en/rest/activity/events). While we want to
avoid using the API to get information regarding every single commit
made to LLVM, we can narrow our calls down to only commits that don't
have any pull request data available via BigQuery. From those calls, we
can determine if a "push" commit actually does have a pull request and,
if it does, whether or not it has been approved.