[CI] Use GraphQL API instead of BigQuery to get review data (#525)
As we are already making calls to the GitHub GraphQL API for data
validation, we can just remove the added complexity of using GitHub
Archive BigQuery as a data source and query the API directly. Using
BigQuery has the advantage of not being rate-limited, but we often have
to query for 50-70 commits via the API anyway due to missing records of
events in GitHub Archive. With more than half of the BigQuery data
points needing amending, it makes more sense to use the API as the
original data source.