[CI] Add dispatch_job script
This patch adds the dispatch_job python script. This script is designed to be
invoked from within a llvm-zorg AnnotatedBuilder and spawn jobs inside of
kubernetes pods on the premerge cluster. This is not directly integrated
into an AnnotatedBuilder given the additional dependencies that we have
on the kubernetes client API.

Reviewers: lnihlen, dschuff, Keenuts, gburgessiv, cmtice

Reviewed By: cmtice

Pull Request: https://github.com/llvm/llvm-zorg/pull/526