[CI] Add workflow to build premerge-buildbot container
This patch sets up a workflow in a similar vein to all the other container
images for building the premerge-buildbot container and publishing it to
GHCR. There is a lot of duplicate code and eventually this should be
refactored into an action. This is on my TODO list.

Reviewers: cmtice

Reviewed By: cmtice

Pull Request: https://github.com/llvm/llvm-zorg/pull/528