[CI] Export scraped commit data to a BigQuery dataset (#532)
Currently, the data we scrape and process regarding LLVM commits isn't
persistent and cannot be referenced outside of each CronJob invocation.
This change uploads scraped and parsed LLVM commit data to a new
BigQuery dataset, so that we may access and reuse this data without
having to requery and reparse the same commits to llvm-project.