[GreenDragon] Duplicate and templatize clang-stage2-Rthinlto for bisection
The new approach modularizes and templatizes the job so that we can
automatically bisect build failures down to the commit responsible for them. The
new jobs written in this style use the name-inspecting symlink trick that clang
itself uses to determine driver mode, for organizational simplicity and
deduplication of configuration. For now, I am duplicating the job wholesale,
until we're comfortable with replacing all of the existing clang-* jobs with
this style.

As part of this, we're adding a templatized bisection job and corresponding
runner. Currently, these cooperate to bisect the clang-stage2-Rthinlto-v2 job's
failures, but going forward we should be set up to re-use that for any other
job's bisection needs. We will also be well set up to add a test bisection
runner, which could leverage the artifacts saved to S3 as a caching mechanism to
speed up the search.