[CI] Stop counting commit authors are reviewers (#545)
When a pull request author comments on their own pull request, it counts
as a review and is captured when we query the GraphQL API for reviewers
of a pull request. We shouldn't be counting self-reviews when collecting
LLVM metrics, so this change removes the author of a commit from the set
of reviewers.