[CI] Double RAM in Service Pool (#546)
The operational metrics CronJob was having trouble scheduling due to
memory pressure. Change from highcpu instances to standard instances
which increases the amount of RAM from 4GB to 16GB per instance. This
costs slightly more, but the difference is negligible. If we end up
being that concerned, we can use custom sized instances and move to only
8GB of RAM, but for now this should be fine.