[CI] Bump libc++ container version
https://github.com/llvm/llvm-project/commit/4fd41c4afbc76ead0c46e80990f616d21dd983f6
updated the runner version in the container. Bump to that to take advantage of the
runner vesion update. Bump both the next runner set and the existing runner set
given we are currently running all jobs on the next runner set. This is safe to bump
given this was a runner only version bump.