Add DebugifyBuilder for debug location coverage testing (#493)
This patch adds a new build factory for running tests using Debugify (see
"How To Update Debug Info" in the LLVM docs[0]) for the purposes of
detecting debug info errors as proposed on Discourse[1]. This builder is
very similar to the `TestSuiteBuilder`, but it adds some required CMake
flags to the LLVM and Test Suite builds, and adds an extra step where we
use a script within `llvm/utils` to evaluate the output of Debugify.

As part of implementing this, I had to make some small changes to the
`TestSuiteBuilder` to allow CMake flags to be passed for the Test Suite
build, as currently it only accepts flags for the LLVM build. These
changes should be a no-op for all existing builds; only by passing the
new `extra_test_suite_configure_args` parameter should this have any
effect.

[0] https://llvm.org/docs/HowToUpdateDebugInfo.html#test-original-debug-info-preservation-in-optimizations
[1] https://discourse.llvm.org/t/rfc-require-real-or-annotated-source-locations-on-all-instructions/86816